[skip ci] mysql ignores null order (still present in the schema)
