+++ server/src-lib/Hasura/RQL/DDL/Schema/Cache/Dependencies.hs
@@ -15,9 +15,7 @@ import           Data.Monoid                        (First)
 
 import           Hasura.RQL.DDL.Schema.Cache.Common
 import           Hasura.RQL.Types
-import           Hasura.SQL.Builder
 import           Hasura.SQL.Text
-import           Hasura.SQL.Types
 
 -- | Processes collected 'CIDependency' values into a 'DepMap', performing integrity checking to
 -- ensure the dependencies actually exist. If a dependency is missing, its transitive dependents are
