+++ server/src-lib/Hasura/RQL/DDL/Schema/Cache/Permission.hs
@@ -22,7 +22,6 @@ import           Hasura.RQL.DDL.Schema.Cache.Common
 import           Hasura.RQL.Types
 import           Hasura.RQL.Types.Catalog
 import           Hasura.Session
-import           Hasura.SQL.Builder
 import           Hasura.SQL.Text
 import           Hasura.SQL.Types
