+++ server/src-lib/Hasura/RQL/DDL/Schema/Table.hs
@@ -54,7 +54,6 @@ import           Hasura.RQL.DDL.Schema.Rename
 import           Hasura.RQL.Types
 import           Hasura.RQL.Types.Catalog
 import           Hasura.Server.Utils
-import           Hasura.SQL.Builder
 import           Hasura.SQL.Text
 import           Hasura.SQL.Types
